+++ mobile/android/fenix/app/src/test/java/org/mozilla/fenix/ext/ActivityTest.kt
@@ -5,9 +5,11 @@
 package org.mozilla.fenix.ext
 
 import android.app.Activity
+import android.content.Intent
 import android.view.View
 import android.view.WindowManager
 import io.mockk.mockk
+import mozilla.components.support.utils.EXTRA_ACTIVITY_REFERRER_PACKAGE
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertThrows
 import org.junit.Assert.assertTrue
@@ -75,5 +77,22 @@ class ActivityTest {
         val activity = mockk<Activity>()
 
         assertThrows(IllegalArgumentException::class.java) { activity.getBreadcrumbMessage(mockk()) }
+    }
+
+    @Test
+    fun `WHEN an intent is from within the app THEN return true`() {
+        // We need to create a dummy activity so that we can get the right packageName in the
+        // test environment so that we can put it into a new activity to verify the extra is set.
+        // We don't hardcode a packageName in here because the value changes based on the variant
+        // that we run these from (e.g. debug, beta)
+        val dummyActivity = Robolectric.buildActivity(Activity::class.java).create().get()
+        val intent = Intent().apply {
+            putExtra(EXTRA_ACTIVITY_REFERRER_PACKAGE, dummyActivity.packageName)
+        }
+        val activity = Robolectric.buildActivity(Activity::class.java, intent)
+            .create()
+            .get()
+
+        assertTrue(activity.isIntentInternal())
     }
 }